moved foreground tests to their own file
